server/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers
django-filter==23.3
Pillow
psycopg2-binary==2.9.9  # Removed due to Python 3.13 compatibility issues
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0 